[fix] Fix x-axis ticks formatting for absolute time alignment
- replace `getFormattedYAxis` function content in HighPlot with dynamic `formatYAxisByDefault` function
